Conflicts:
	src/ziozio/service/impl/ClothServiceImpl.java

+++ src/ziozio/dto/Location.java
@@ -2,16 +2,16 @@ package ziozio.dto;
 
 public class Location implements DTO {
 	
-	private int location_code;
+	private long location_code;
 	private String location_name;
 	@Override
 	public String toString() {
 		return "Location [location_code=" + location_code + ", location_name=" + location_name + "]";
 	}
-	public int getLocation_code() {
+	public long getLocation_code() {
 		return location_code;
 	}
-	public void setLocation_code(int location_code) {
+	public void setLocation_code(long location_code) {
 		this.location_code = location_code;
 	}
 	public String getLocation_name() {
